#288 refactor unit test
